+++ lib/src/base/mega_portrait_layout.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/cupertino.dart' show CupertinoPageRoute;
 import 'package:flutter/material.dart' show Icons;
-import 'package:flutter/widgets.dart';
 import 'package:mega/mega.dart';
 
 class MegaPortraitLayout extends StatefulWidget {